explicit length with fill bytes

+++ dadl-processor/src/main/java/org/ops4j/dadl/io/BitStreamReader.java
@@ -180,7 +180,7 @@ public abstract class BitStreamReader extends ImageInputStreamImpl {
         return result;
     }
 
-    public void skipBits(int bitCnt) throws IOException {
+    public void skipBits(long bitCnt) throws IOException {
         setBitPosition(getBitPosition() + bitCnt);
     }
